factor: revert f32 demotion of forward log returns — restores +0.0031 baseline
The "f32 throughout — no long-horizon accumulation" optimization in
3002e8d's `forward_log_returns` (data.py) and `precompute_inputs`
`daily_log_ret`/`log_p` (train.py) drifted the linear-arm mean val IC
from +0.0031 to +0.0005 on the SSL walkforward — a 6× regression.

Mechanism: `pearson_rank_ic` is plain cross-sectional Pearson on raw
forward log returns, not on ranks. Its covariance numerator
`sum((scores - s_mean) * (fwd_returns - r_mean) * mask)` cancels
catastrophically when val IC is small (~0.003). f32 forward returns
introduce ulp-scale precision loss at exactly the magnitude where the
signal lives — `log_p` at magnitude ~6 has f32 ulp ~7e-7, which
propagates through the subtraction into the IC numerator.

f64-then-cast-to-f32 at the Tensor boundary preserves ~200× more
precision than f32-throughout because the final cast quantizes to
~3e-9 ulp at the result magnitude (~0.05) rather than carrying the
~7e-7 input rounding through.

Verified: linear arm with f64 forward log returns reproduces the doc's
per-window numbers bit-for-bit (−0.000, +0.001, +0.003, −0.002, +0.010,
+0.006 → mean +0.0031). Same backbone, same universe, same
hyperparameters.

The (D, N) panel is small (~78 MB at full pool); the f64-vs-f32 byte
savings aren't worth the SNR loss on a metric that lives in the
cancellation regime.

+++ apps/factor/src/factor/data.py
@@ -203,13 +203,17 @@ def forward_log_returns(
     to close-of-(i+rebal_days). The trailing `rebal_days` rows are NaN
     (the future window doesn't fit).
     """
-    # f32 throughout — log() of f64 prices then subtracted in-window
-    # has no accumulation, so f32 precision (~7 digits) is plenty for
-    # forward log returns on the [-1, 1] range typical of horizons we
-    # use. Keeps a `(D, N)` panel at half the bytes.
-    log_p = np.log(np.maximum(prices, 1e-12)).astype(np.float32)
+    # f64 throughout — pearson_rank_ic's covariance numerator is the
+    # difference of two near-equal sums of (score - mean) * (fwd - mean),
+    # which cancels catastrophically when val IC is small (~0.003 here).
+    # f32 forward returns introduce ulp-scale precision loss at exactly
+    # the magnitude where the signal lives, drifting val IC by O(1e-3).
+    # The (D, N) panel is small (~78 MB at full pool); the f64-vs-f32
+    # bytes saved aren't worth the SNR loss. Caller demotes to f32 at
+    # the Tensor boundary in precompute_inputs.
+    log_p = np.log(np.maximum(prices, 1e-12))
     D, N = prices.shape
-    fwd = np.full((D, N), np.nan, dtype=np.float32)
+    fwd = np.full((D, N), np.nan, dtype=np.float64)
     if D > rebal_days:
         fwd[:D - rebal_days] = log_p[rebal_days:] - log_p[:D - rebal_days]
     return fwd
